适配 BBS FS 2.7：修复 tick 浮点化、overlay 拆分、移除失效 reset 重定向、新增 UnknownFormRenderer
- gradle.properties: BBS jar 指向 2.7 (1.20.1/1.20.4)
- fabric.mod.json: 依赖 bbs >=2.7-0
- UIEquipmentKeyframeTransform: getAutoKeyframeTick Integer->Float, getAutoKeyTransform(int)->(float)
- ModelFormCMLMixin: recordingPoseTransformOverlays 拆为 recordingPoseOverlays+recordingTransformOverlays, CML overlay 数量取二者之和
- UIReplaysEditorUtilsMixin: fillAnimationPose 末参 int current -> float current
- UIPoseTransformsBoneSkipMixin: 删除 2.7 中目标调用已不存在的 reset @Redirect，清理无用 Consumer import
- BBSFSloveCMLClient: 注册 UnknownFormRenderer 空渲染器防止缺失 form 类型 NPE
- .gitignore: 添加 .error_report_analysis/

+++ src/main/java/gbeic/bbsplusplus/client/keyframes/UIEquipmentKeyframeTransform.java
@@ -45,7 +45,7 @@ public class UIEquipmentKeyframeTransform extends UIKeyframePropTransform
     @Override
     protected void applyToSelection(Consumer<Transform> consumer)
     {
-        Integer autoTick = this.editor == null ? null : this.editor.getAutoKeyframeTick();
+        Float autoTick = this.editor == null ? null : this.editor.getAutoKeyframeTick();
 
         if (autoTick != null)
         {
@@ -84,7 +84,7 @@ public class UIEquipmentKeyframeTransform extends UIKeyframePropTransform
     }
 
     @Override
-    protected Transform getAutoKeyTransform(int tick)
+    protected Transform getAutoKeyTransform(float tick)
     {
         UIKeyframeSheet sheet = this.editor.getGraph().getSheet(this.keyframe);
         Keyframe<?> recorded = this.ensureEquipmentKeyframe(sheet, tick);
@@ -116,7 +116,7 @@ public class UIEquipmentKeyframeTransform extends UIKeyframePropTransform
         }
     }
 
-    private Keyframe<?> ensureEquipmentKeyframe(UIKeyframeSheet sheet, int tick)
+    private Keyframe<?> ensureEquipmentKeyframe(UIKeyframeSheet sheet, float tick)
     {
         if (sheet == null || !EquipmentKeyframeTransforms.isEquipmentChannel(sheet.channel.getId()))
         {
